jiuwenclaw/agentserver/tools/browser-move/pyproject.toml
[project]
name = "browser-move"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "authlib==1.6.7",
    "beartype==0.22.9",
    "cacheout==0.16.0",
    "cachetools==7.0.1",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "cfgv==3.5.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cloudpickle==3.1.2",
    "codecov==2.1.13",
    "colorama==0.4.6",
    "coverage==7.13.4",
    "croniter==6.0.0",
    "cryptography==46.0.5",
    "cyclopts==4.5.2",
    "dashscope==1.25.12",
    "diskcache==5.6.3",
    "distlib==0.4.0",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.22.4",
    "docx2txt==0.9",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.1",
    "fakeredis==2.33.0",
    "fastmcp==2.14.5",
    "filelock==3.20.3",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "greenlet==3.3.1",
    "grpcio==1.78.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.4.1",
    "identify==2.6.16",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "iniconfig==2.3.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.1.0",
    "jaraco-functools==4.4.0",
    "jiter==0.13.0",
    "json-repair==0.57.1",
    "jsonref==1.1.0",
    "jsonschema==4.26.0",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "keyring==25.7.0",
    "lupa==2.6",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "mcp==1.26.0",
    "mdurl==0.1.2",
    "mermaid-py==0.8.3",
    "more-itertools==10.8.0",
    "multidict==6.7.1",
    "nodeenv==1.10.0",
    "numpy==2.4.2",
    "oauthlib==3.3.1",
    "openai==2.20.0",
    "openapi-pydantic==0.5.1",
    "openjiuwen==0.1.7",
    "opentelemetry-api==1.39.1",
    "orjson==3.11.7",
    "packaging==26.0",
    "pandas==3.0.0",
    "pathable==0.4.4",
    "pathvalidate==3.3.1",
    "pdfminer-six==20251230",
    "pdfplumber==0.11.9",
    "pillow==12.1.1",
    "platformdirs==4.5.1",
    "pluggy==1.6.0",
    "portalocker==3.2.0",
    "pre-commit==4.5.1",
    "prometheus-client==0.24.1",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "py-key-value-aio==0.3.0",
    "py-key-value-shared==0.3.0",
    "pycparser==3.0",
    "pycryptodome==3.23.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pydocket==0.17.7",
    "pygments==2.19.2",
    "pyjwt==2.11.0",
    "pymilvus==2.6.9",
    "pypdfium2==5.4.0",
    "pyperclip==1.11.0",
    "pysbd==0.3.4",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "python-json-logger==4.0.0",
    "python-multipart==0.0.22",
    "pytz==2025.2",
    "pywin32==311",
    "pywin32-ctypes==0.2.3",
    "pyyaml==6.0.3",
    "redis==7.1.1",
    "referencing==0.36.2",
    "regex==2026.1.15",
    "requests==2.32.5",
    "rich==14.3.2",
    "rich-rst==1.3.2",
    "rpds-py==0.30.0",
    "ruff==0.15.0",
    "safetensors==0.7.0",
    "setuptools==82.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==2.0.46",
    "sse-starlette==3.2.0",
    "starlette==0.52.1",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "tqdm==4.67.3",
    "transformers==5.1.0",
    "typer==0.23.0",
    "typer-slim==0.23.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uvicorn==0.40.0",
    "virtualenv==20.36.1",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
]
